User controller manage exceptions

+++ src/main/java/org/ieti/proyecto/controller/user/UserController.java
@@ -1,5 +1,6 @@
 package org.ieti.proyecto.controller.user;
 
+import org.ieti.proyecto.exception.UserNotFoundException;
 import org.ieti.proyecto.models.users.User;
 import org.ieti.proyecto.models.users.UserDTO;
 import org.ieti.proyecto.service.user.UserService;
@@ -9,6 +10,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController("/v1/api/user")
 public class UserController {
@@ -27,16 +29,29 @@ public class UserController {
 
     @GetMapping("/{id}")
     public ResponseEntity<User> getUserById(@PathVariable("id") String userId) {
-        return new ResponseEntity<>(userService.findById(userId).get(), HttpStatus.OK);
+        Optional<User> userSearched = userService.findById(userId);
+        if (userSearched.isEmpty()) {
+            throw new UserNotFoundException(userId);
+        }
+
+        return new ResponseEntity<>(userSearched.get(), HttpStatus.OK);
     }
 
     @PutMapping("/{id}")
     public ResponseEntity<User> updateUser(@RequestBody UserDTO userUpdates, @PathVariable("id") String userId) {
+        if (userService.findById(userId).isEmpty()) {
+            throw new UserNotFoundException(userId);
+        }
+
         return new ResponseEntity<>(userService.update(userUpdates, userId), HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteUser(@PathVariable("id") String userId) {
+        if (userService.findById(userId).isEmpty()) {
+            throw new UserNotFoundException(userId);
+        }
+
         userService.deleteById(userId);
         return new ResponseEntity<>(HttpStatus.OK);
     }
